Some other bits - few tweaks to make writing to the DB work.

+++ server/src/database/migrations/20251027061056_add-esp32-externalAddresses.ts
@@ -1,40 +1,42 @@
 import type { Knex } from "knex";
 
 export async function up(knex: Knex): Promise<void> {
-  await knex.schema.createTableIfNotExists("external_devices", (table) => {
-    table.increments("id").notNullable();
-    table.string("name").notNullable();
-    table.string("type").notNullable();
-    table.string("externalAddress").notNullable();
-    table.string("secureToken").defaultTo(null);
-    table.primary(["id"]);
-  });
+  if (!(await knex.schema.hasTable("external_devices"))) {
+    await knex.schema.createTable("external_devices", (table) => {
+      table.increments("id").notNullable();
+      table.string("name").notNullable();
+      table.string("type").notNullable();
+      table.string("address").notNullable();
+      table.string("secureToken").defaultTo(null);
+      table.primary(["id"]);
+    });
 
-  await knex.schema.alterTable("sensors", (table) => {
-    table
-      .integer("externalDevice_id")
-      .unsigned()
-      .nullable()
-      .defaultTo(null)
-      .after("model")
-      .references("id")
-      .inTable("external_devices")
-      .onDelete("CASCADE")
-      .onUpdate("CASCADE");
-  });
+    await knex.schema.alterTable("sensors", (table) => {
+      table
+        .integer("externalDevice_id")
+        .unsigned()
+        .nullable()
+        .defaultTo(null)
+        .after("model")
+        .references("id")
+        .inTable("external_devices")
+        .onDelete("CASCADE")
+        .onUpdate("CASCADE");
+    });
 
-  await knex.schema.alterTable("outputs", (table) => {
-    table
-      .integer("externalDevice_id")
-      .unsigned()
-      .nullable()
-      .defaultTo(null)
-      .after("model")
-      .references("id")
-      .inTable("external_devices")
-      .onDelete("CASCADE")
-      .onUpdate("CASCADE");
-  });
+    await knex.schema.alterTable("outputs", (table) => {
+      table
+        .integer("externalDevice_id")
+        .unsigned()
+        .nullable()
+        .defaultTo(null)
+        .after("model")
+        .references("id")
+        .inTable("external_devices")
+        .onDelete("CASCADE")
+        .onUpdate("CASCADE");
+    });
+  }
 }
 
 export async function down(knex: Knex): Promise<void> {
